+++ packages/fluentui/accessibility/src/behaviors/Button/buttonBehavior.ts
@@ -1,4 +1,4 @@
-import { keyboardKey, SpacebarKey } from '@fluentui/keyboard-key';
+import { keyboardKey, SpacebarKey } from '../../keyboard-key';
 import { Accessibility, AccessibilityDefinition } from '../../types';
 
 export const buttonBehavior: Accessibility<ButtonBehaviorProps> = props => {