+++ packages/fluentui/accessibility/src/behaviors/Radio/radioGroupItemBehavior.ts
@@ -1,5 +1,5 @@
 import { Accessibility } from '../../types';
-import { SpacebarKey } from '@fluentui/keyboard-key';
+import { SpacebarKey } from '../../keyboard-key';
 import { IS_FOCUSABLE_ATTRIBUTE } from '../../attributes';
 
 /**